Basic Working Frontend
The design is mostly done, some tweaks need to be made. The forms need to have a cancel button or a way to hide the form. The emergency access will need to be reworked, in terms of how it works, so that the user can choose which patient it applies to.