Reworked Emergency Access
Emergency Access has been reworked to be on the individual patient forms instead of a button on the sidebar. I didn't delete the button from the sidebar, only removed it, since if a user does not have access to any ePHI, such as a guest, but needs to access it, then a button outside the form should exist.