Removed "Full Chart" data field
There was an additional patient data field called "Full Chart" that appeared in the patient information, probably created and inserted by the AI when I first generated the project. This update removes it, but further look at the codebase to determine if it's been thoroughly removed from the application may be needed. As of now, it seems to be gone.