Removed the filter
Filter removed for multiple reasons: 1. I couldn't get the filter to work, it might require a full refactor of how the tabs work; 2. It might not be the best way to do it, since the user might need access to information they are allowed, so don't need emergency access, but they didn't request at first.